chore(server): core IoC #50 - getTokenAppInfoFactory
