+++ packages/server/modules/auth/repositories/apps.ts
@@ -13,6 +13,7 @@ import {
   GetApp,
   GetAuthorizationCode,
   GetRefreshToken,
+  GetTokenAppInfo,
   RegisterDefaultApp,
   RevokeExistingAppCredentials,
   RevokeExistingAppCredentialsForUser,
@@ -411,4 +412,32 @@ export const getRefreshTokenFactory =
   (deps: { db: Knex }): GetRefreshToken =>
   async ({ id }) => {
     return await tables.refreshTokens(deps.db).select('*').where({ id }).first()
+  }
+
+export const getTokenAppInfoFactory =
+  (deps: { db: Knex }): GetTokenAppInfo =>
+  async (params: { token: string; appId?: string }) => {
+    const { token, appId } = params
+    const tokenId = token.slice(0, 10)
+
+    const q = tables
+      .apiTokens(deps.db)
+      .select<ServerAppRecord[]>(ServerApps.cols)
+      .where({
+        [ApiTokens.col.id]: tokenId,
+        ...(appId
+          ? {
+              [UserServerAppTokens.col.appId]: appId
+            }
+          : {})
+      })
+      .innerJoin(
+        UserServerAppTokens.name,
+        ApiTokens.col.id,
+        UserServerAppTokens.col.tokenId
+      )
+      .innerJoin(ServerApps.name, ServerApps.col.id, UserServerAppTokens.col.appId)
+      .first()
+
+    return await q
   }
